[Chore #163023304] Create middleware validation helper function
